Incorporate SharedTunnelOptions.swift into Shared.swift
Some contents of SharedTunnelOptions.swift are going to be used in both
OpenVPN and WireGuard, in both iOS and macOS.
